Fix SLS error output on conformity details per minion
On the light theme an sls error is not shown because the font color is
also black. Because it's an error the font color should be red and I
also recommend to show all errors logged by the state.

Using the existing method for the nested_output and adding color support
if a string is used.